cloud/crypto-pipeline/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-storage==3.4.0
requests==2.32.3
matplotlib